Create Two_Sum.cpp
This code defines a class called Solution that contains a public member function called twoSum. The twoSum function takes in a reference to a vector of integers called nums and an integer called target. It returns a vector of integers.

The function begins by declaring an empty vector called ans that will be used to store the indices of the elements in the input vector that sum to the target value.

Next, the function has a nested loop structure. The outer loop iterates over the elements of the nums vector, and the inner loop iterates over the remaining elements of the nums vector, starting from the element after the current element of the outer loop.

Inside the inner loop, the function checks if the elements at indices i and j sum to the target value. If they do, the function adds the indices to the ans vector and breaks out of the inner loop.

Finally, the function returns the ans vector.