Create PalindromeNumber(general_way).cpp
The isPalindrome function takes an integer x as input and returns a boolean value indicating whether x is a palindrome.

The function begins by initializing a result variable result to 0. This variable will be used to store the reverse of x as we iterate through its digits.

Next, the function checks if x is negative or ends with 0 (but is not 0 itself). If either of these conditions is true, the function returns false immediately, because a negative number or a number ending in 0 cannot be a palindrome.

If x passes these checks, the function enters a loop that iterates until x is greater than result. On each iteration of the loop, the following steps are performed:

The last digit of x is extracted using the modulo operator (%). This is done by taking the remainder of x when divided by 10, which will always be the last digit of x.
The last digit of x is removed by dividing x by 10.
The extracted digit is added to the result by multiplying result by 10 and adding the extracted digit. This shifts all the digits in result one place to the left and adds the extracted digit to the end.
After the loop completes, the function checks whether x is equal to result, or whether x is equal to result with the last digit removed (for odd-length palindromes). If either of these conditions is true, the function returns true to indicate that x is a palindrome. Otherwise, it returns false.