Create RomanToInteger.cpp
The above code defines a class called Solution that contains a single method romanToInt, which converts a Roman numeral in string form to an integer.

The romanToInt function first declares a few variables:

result: this variable will hold the integer value of the Roman numeral. It is initialized to 0.
sFirst: this variable will hold the integer value of the current character in the input string.
sSecond: this variable will hold the integer value of the next character in the input string.
The function then iterates through the characters of the input string with a for loop. For each character, it first determines its integer value using the value function and stores it in sFirst. It then determines the integer value of the next character (if it exists) and stores it in sSecond.

Next, the function compares the values of sFirst and sSecond. If sFirst is greater than or equal to sSecond, it means that the current character does not represent a "subtractive" Roman numeral, and its value can simply be added to the result. On the other hand, if sFirst is less than sSecond, it means that the current character represents a "subtractive" Roman numeral, and its value should be subtracted from the value of the next character instead. In this case, the loop counter is incremented by one to skip the next character.

Finally, the function returns the integer value of the Roman numeral stored in result.