Create longestCommonPrefix.cpp
This code is finding the longest common prefix among a list of strings stored in a vector called strs.

The first line sorts the strings in the vector in ascending order. This is done so that the first and last strings in the vector can be compared to find the longest common prefix among all of the strings.

The second line initializes an empty string called answer to store the longest common prefix.

The third and fourth lines store the first and last strings in the vector in separate variables called a and b, respectively.

The fifth line begins a loop that will iterate through the characters of the first string, a. The loop continues until it has reached the end of the string, which is determined by the loop control variable i being less than the length of a.

Inside the loop, the code compares the characters at the current index in both a and b. If the characters are the same, the character is added to the answer string. If the characters are different, the loop is broken and the function moves on to the next step.

Finally,  returns the answer string as the result.