Create validParentheses.cpp
This code is a solution to the problem of checking whether a string of brackets is valid or not.

The function takes a string of brackets as input and returns a boolean value indicating whether the brackets are balanced and properly nested.

Here is how the function works:

The function creates an empty stack of characters called "st". This stack will be used to store the brackets as they are encountered in the string.

The function then iterates through each character in the string.

If the stack is empty or the character is an opening bracket ('(', '[', or '{'), it pushes the character onto the stack.

If the character is a closing bracket (')', ']', or '}'), the function checks the top element of the stack. If the top element does not match the corresponding opening bracket for the closing bracket, it returns false. For example, if the top element is '[' and the current character is '}', the function returns false because these brackets do not match.

If the top element does match the corresponding opening bracket, the function pops the top element off the stack. This indicates that a pair of brackets has been matched and can be removed from the stack.

After the function has finished iterating through the string, it returns true if the stack is empty (all brackets have been matched) or false if the stack is not empty (not all brackets have been matched).