Create firstOccurrenceIndex.cpp
This code is a solution to the problem of finding the index at which a needle string appears in a haystack string.

The function takes in two strings as input: haystack and needle. It first stores the lengths of these strings in the variables n and m respectively.

It then enters a loop that iterates through the haystack string. The loop variable i goes from 0 to n-m, which means it will consider substrings of haystack that are as long as needle.

Inside this loop, it initializes a second loop variable j to 0. It then enters a second loop that iterates through the needle string.

Inside this second loop, it comparesthe j-th character of needle to the (i+j)-th character of haystack. If these characters are not equal, it breaks out of the loop.

If the loop completes without breaking, it means that the needle has been found in the haystack starting from index i. In this case, the function returns i as the output.

If the needle was not found in the haystack, the function returns -1.