Create bookBox.cpp
Here is an explanation of the code:

The program is written in C++ and implements a simple stack using an array. It allows the user to perform the following actions:

Insert a book into the stack
Remove a book from the stack
Display the books in the stack
Exit the program
The program runs in a loop, prompting the user to choose an option between 1 and 4. The user's choice is then handled using a switch statement.

The case 1: block of code is executed when the user selects option 1 (insert a book). The program checks if the stack is full by checking the value of the top variable. If the stack is full, it displays an error message. If the stack is not full, the program prompts the user to enter a book number and then adds the book to the top of the stack using the strcpy function from the string.h library.

The case 2: block of code is executed when the user selects option 2 (remove a book). The program checks if the stack is empty by checking the value of the top variable. If the stack is empty, it displays an error message. If the stack is not empty, the program prompts the user to enter a book number and then removes the book at the top of the stack by decrementing the top variable.

The case 3: block of code is executed when the user selects option 3 (display the books in the stack). The program checks if the stack is empty by checking the value of the top variable. If the stack is empty, it displays an error message. If the stack is not empty, it displays the books in the stack by iterating through the array from the top of the stack down to the bottom.

The case 4: block of code is executed when the user selects option 4 (exit the program). It simply exits the program by returning 0.

The default: block of code is executed when the user enters an invalid choice (a number other than 1, 2, 3, or 4). It displays an error message.

The program runs in a loop, so it will continue to prompt the user for a choice until they select option 4 (exit the program).