Create NextPermutation.cpp
The function nextPermutation takes a vector of integers nums and modifies it to the next lexicographically larger permutation.

The algorithm works as follows:

Find the first element from the right that is smaller than its right neighbor. Let's call this element nums[i]. If no such element is found, the permutation is already at its maximum, so we reverse the whole vector to get the minimum permutation.

Otherwise, we proceed to modify the permutation. Find the first element from the right that is larger than nums[i]. Let's call this element nums[k].

Swap nums[i] and nums[k].

Reverse the elements to the right of i to get the next permutation.