Create removeElement.cpp
his code defines a class called Solution that contains a member function called removeElement. This function takes a reference to a vector of integers nums and an integer val as input and returns an integer.

The function first initializes a variable i to zero. It then iterates through the elements of the vector nums using a for loop. For each element nums[j] in the vector, if it is not equal to the given value val, it is copied to the element at index i and i is incremented.

At the end of the loop, the function returns the value of i, which is the new length of the modified vector. The modified vector will contain all the elements of the original vector except for the ones that are equal to val.