Create circularLinkedList.cpp
This code implements a circular linked list in C++. A circular linked list is a data structure in which each node has a next pointer that points to the next node in the list, and the last node in the list has a next pointer that points back to the first node in the list, forming a circular structure.

The node class represents a node in the circular linked list. It has two member variables: data, which stores the data stored in the node, and next, which is a pointer to the next node in the list. The node class has a constructor, which initializes the data and next variables when a new node is created, and a destructor, which frees the memory for a node when it is deleted.

The insertnode function takes in a reference to the tail of the list, the data of the node after which a new node is to be inserted, and the data for the new node. If the list is empty, it creates a new node and sets it as the tail. Otherwise, it searches for the node with the given data and inserts a new node after it.

The print function takes in the tail of the list and prints the data of all nodes in the list.

In the main function, the tail of the list is initialized to NULL, and several nodes are inserted into the list using the insertnode function. The list is then printed using the print function.