Create DivideTwoIntegers.cpp
This is a C++ implementation of an algorithm for performing integer division. The algorithm works by continuously subtracting the divisor from the dividend until the dividend is less than the divisor. The number of times the divisor was subtracted from the dividend is the quotient of the division.

The function takes in two integer arguments: dividend and divisor, which represent the dividend and divisor of the division operation, respectively. It returns an integer value representing the result of the division.

The first step in the function is to handle two special cases:

If the dividend is the minimum value that an integer can take on (INT_MIN) and the divisor is -1, the result of the division would be greater than the maximum value that an integer can take on (INT_MAX). Therefore, we return INT_MAX in this case.
If the dividend is INT_MIN and the divisor is 1, the result would be INT_MIN. Therefore, we return INT_MIN in this case.
Next, the function determines the sign of the result by exclusive ORing the signs of the dividend and divisor. If the signs are different, the result will be negative. Otherwise, it will be positive.

The function then takes the absolute values of the dividend and divisor to make the division process easier. It stores the absolute value of the dividend in a variable called D and the absolute value of the divisor in a variable called d.

The main division loop starts by checking if D (the absolute value of the dividend) is greater than or equal to d (the absolute value of the divisor). If it is, the loop continues. If it is not, the division is complete and the function can return the result.

Inside the loop, the function stores the value of d in a temporary variable called temp. It also initializes a variable called quotient to 1.

The function then enters a nested loop where it continuously doubles the value of temp and quotient until temp is greater than D. It then adds half the value of quotient to the result (stored in the variable ans) and subtracts half the value of temp from D.

This process continues until the outer loop condition (D being greater than or equal to d) is no longer true. At this point, the function returns the result (stored in ans) multiplied by the sign determined earlier