Create plusOne.cpp
The above code defines a class Solution with a single function plusOne, which takes a reference to a vector of integers as input and returns a vector of integers. The purpose of this function is to add one to the input vector, which is interpreted as a number in base 10.

The function starts by looping through the elements of the input vector from the last element to the first. At each iteration, it checks if the current element is 9. If it is not, it simply adds 1 to the element and returns the modified vector. If the current element is 9, it sets it to 0 and continues iterating.

If the loop finishes and all the elements of the vector are 9, it means that we need to add a leading 1 to the vector to represent the carryover from adding one. For example, if the input vector is [9, 9, 9], the function would return [1, 0, 0, 0].

Finally, the function returns the modified vector.