Create addBinary.cpp
The addBinary function takes in two strings, a and b, which represent binary numbers. The function returns a string representing the sum of the two binary numbers.

First, the function initializes two variables, i and j, to store the lengths of the input strings a and b, respectively. The indices of the characters in the strings are used to loop through the strings from right to left (i.e., from the least significant digit to the most significant digit). The carry variable is initialized to 0 and is used to store any carry value that occurs when adding two digits. The res string is initialized to an empty string and will be used to store the result of the addition.

The function then enters a loop that continues until both input strings have been fully processed. Inside the loop, the sum variable is initialized to the value of carry. If the i index is still within the bounds of the a string, the value at the i index is added to sum. Similarly, if the j index is still within the bounds of the b string, the value at the j index is added to sum.

The carry value is then updated based on the value of sum. If sum is greater than 1, it means that a carry occurred during the addition, and the carry value is set to 1. Otherwise, the carry value is set to 0.

The least significant digit of the sum is then added to the res string using the to_string function. The i and j indices are decremented to move on to the next least significant digits in the input strings.

After the loop finishes, if the carry value is 1, it means that there is a final carry value that needs to be added to the result. This carry value is added to the res string using the to_string function.

Finally, the res string is reversed using the reverse function and returned