Create removeDuplicateFromSortedList.cpp
This code is a solution for a problem where you are given a singly linked list, and you need to remove all the nodes in the list that have duplicate values.

The deleteDuplicates function takes in a pointer to the head of the list (called head) and returns a pointer to the head of the list with all the duplicate nodes removed.

The function sets up a pointer called curr that points to the head of the list. Then, it enters a loop that continues as long as curr is not nullptr (meaning the end of the list has not been reached) and curr has a next node.

Inside the loop, if the value of the current node (curr->val) is the same as the value of the next node (curr->next->val), then the next node is removed from the list by setting the next node of the current node to be the one after the next node (curr->next = curr->next->next). Otherwise, the loop moves on to the next node by setting curr to be curr->next.

Finally, the function returns the head of the list.