Create deleteColumnsToMakeSorted.cpp
The minDeletionSize function takes in a single parameter, strs, which is a vector of strings. The function returns an integer representing the minimum number of columns that need to be deleted in order for the strings in the strs vector to be lexicographically sorted (i.e., in alphabetical order).

The function first initializes a counter cnt to zero. This counter will be used to keep track of the number of columns that need to be deleted.

The function then calculates the number of rows (i.e., the number of strings) and the number of columns (i.e., the length of each string) in the strs vector. These values are stored in the variables row and col, respectively.

The function then enters a nested loop structure, where the outer loop iterates through each column (j) in the strings, and the inner loop iterates through each row (i) in the strings (except the last one).

Inside the inner loop, the function compares the character at the current position (j) in the current row (i) with the character at the same position in the next row (i+1). If the character in the current row is greater than the character in the next row, this means that the strings are not sorted in alphabetical order at this point. In this case, the function increments the counter cnt by 1 and breaks out of the inner loop using the break statement. This is because once we find that a column needs to be deleted, there is no need to continue checking the rest of the rows.

After the nested loops have completed, the function returns the value of cnt, which is the number of columns that need to be deleted in order for the strings to be lexicographically sorted.