Create validPalindrome.cpp
This code defines a class called Solution which has a public member function isPalindrome. This function takes a string as input and returns a boolean value indicating whether or not the string is a palindrome.

A palindrome is a word or phrase that reads the same backwards and forwards. For example, "racecar" and "level" are palindromes.

The function first initializes two pointers called left and right to the beginning and end of the string, respectively. It then enters a loop that continues as long as left is less than right.

Inside the loop, the function checks the characters at the left and right pointers. If either character is not an alphanumeric character (i.e. a letter or a number), the corresponding pointer is incremented or decremented. This allows the function to ignore non-alphanumeric characters when checking for palindromes.

If both characters are alphanumeric and are not equal (ignoring case), the function returns false immediately, since the string cannot be a palindrome. If the characters are equal, the pointers are incremented and decremented, respectively, and the loop continues.

If the loop finishes, it means that the entire string has been processed and all alphanumeric characters were found to be equal, so the function returns true.