Create climbingStairs.cpp
This function returns the number of ways to climb a given number of stairs, n, by taking either 1 or 2 steps at a time.

The function first checks if there are no stairs (n==0), in which case it returns 0 since there are no stairs to climb. If there is only 1 stair (n==1), the function returns 1 since there is only 1 way to climb it (by taking 1 step). If there are 2 stairs (n==2), the function returns 2 since there are 2 ways to climb them: either by taking 1 step at a time or by taking 2 steps at once.

The function then initializes three variables: total_steps, forOneStep, and forTwoStep. total_steps will keep track of the total number of ways to climb the stairs, forOneStep will keep track of the number of ways to climb 1 stair at a time, and forTwoStep will keep track of the number of ways to climb 2 stairs at a time. The initial values of forOneStep and forTwoStep are set to 2 and 1, respectively, to reflect the number of ways to climb the first 2 stairs.

The function then enters a loop that iterates through the remaining stairs, starting from the 3rd stair (for(int i=2;i<n;i++)). On each iteration of the loop, the total number of ways to climb the stairs is updated to be the sum of the number of ways to climb 1 stair and 2 stairs at a time on the previous step (total_steps = forOneStep + forTwoStep). The number of ways to climb 2 stairs at a time is then updated to be the previous number of ways to climb 1 stair at a time (forTwoStep = forOneStep), and the number of ways to climb 1 stair at a time is updated to be the total number of ways for the current step (forOneStep = total_steps).

After the loop has completed, the function returns the total number of ways to climb the stairs, which is stored in the total_steps variable.