Create minimumRoundsToCompleteAllTasks.cpp
 This code is solving the problem of finding the minimum number of rounds required to complete a given set of tasks. Each round consists of three tasks, and each task is represented by an integer. The function takes in a vector of integers called tasks, which represents the set of tasks that need to be completed.

The first step in the function is to create a map called mp that stores the frequency of each task in the tasks vector. This is done using a for loop that iterates through the tasks vector and increments the count of each task in the map.

Next, the function initializes a variable ans to 0. This variable will be used to store the final answer, which is the minimum number of rounds required to complete all of the tasks.

The function then iterates through the map using a for loop. For each task in the map, the function checks if there is only one task of this type. If this is the case, then it is not possible to complete this task because a round consists of three tasks, and we can only complete a task if there are at least three tasks of the same type. In this case, the function returns -1.

If there is more than one task of this type, then the function calculates the minimum number of rounds required to complete these tasks. If the number of tasks is divisible by 3, then the function adds the quotient of this division to the ans variable. If the number of tasks is not divisible by 3, then the function adds the quotient + 1 to the ans variable. This is because any remainder of tasks will require an additional round to complete.

Finally, after iterating through all of the tasks in the map, the function returns the value of ans, which is the minimum number of rounds required to complete all of the tasks.