Create pattern4.cpp
This code defines a class called Solution that contains a function called printTriangle. The function takes in an integer n as input and prints out a triangle of numbers with n rows. The triangle consists of the numbers 1 through n, with each number being repeated in each row.

For example, if n = 3, the triangle would be:
1
2 2
3 3 3

The main function reads in a number of test cases t and, for each test case, reads in an integer n and calls the printTriangle function with n as the input. The function then prints out the triangle as described above. The main function returns 0 at the end.