Create MinimumNumberofArrows.cpp
This is a C++ implementation of a solution to the "Minimum Number of Arrows to Burst Balloons" problem. The problem can be stated as follows: given a list of balloons, each represented by a range (a start point and an end point), find the minimum number of arrows needed to burst all the balloons.

The solution first sorts the list of balloons by their start point. Then it initializes the last burst balloon's end point to the first balloon's end point, and the count of arrows used to 1. It then iterates through the list of balloons, and for each balloon, it compares its start point to the last burst balloon's end point. If the current balloon's start point is greater than the last burst balloon's end point, it increments the arrow count and updates the last burst balloon's end point to the current balloon's end point. If the current balloon's start point is less than or equal to the last burst balloon's end point, it updates the last burst balloon's end point to the minimum of the last burst balloon's end point and the current balloon's end point. This ensures that we use the minimum number of arrows possible. After all balloons have been processed, the function returns the arrow count.