Create isomorphicStrings.cpp
This code is a solution to a problem of checking if two strings are isomorphic.

Isomorphic strings are two strings that can be transformed into each other by replacing certain characters with other characters. However, the rule is that each character in the first string can only be replaced by a character that has not yet been used to replace any other character.

For example, the strings "foo" and "app" are isomorphic because we can transform "foo" into "app" by replacing 'f' with 'a', 'o' with 'p', and 'o' with 'p'. Similarly, we can transform "app" into "foo" by replacing 'a' with 'f', and 'p' with 'o'.

The solution uses two hash maps, map_s and map_t, to store the mapping of characters in the two strings, s and t, respectively. It then iterates through all characters in both strings, and for each character, it checks if the mapping of the character in s is the same as the mapping of the character in t. If it is not, it means that the strings are not isomorphic, and the function returns false. If the loop completes without returning false, it means that the strings are isomorphic, and the function returns true.